requirements.txt
asgiref==3.4.1
astroid==2.8.0
backcall==0.2.0
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
gunicorn==20.1.0
httpie==2.5.0
httpie-jwt-auth==0.4.0
idna==3.2
ipdb==0.13.9
ipython==7.28.0
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.6.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.3.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymongo==3.11.4
PySocks==1.7.1
python-jsonrpc-server==0.3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
requests-toolbelt==0.9.1
snooty-lextudio==1.11.4.dev0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
typing-extensions==3.10.0.2
ujson==1.35
urllib3==1.26.7
watchdog==1.0.2
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.12.1
